requirements copy.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0